requirements.txt
anyio==4.8.0
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
python-telegram-bot==21.11.1
rich==13.9.4
sniffio==1.3.1
typing_extensions==4.12.2
